Troop Selector Menu Integration
Troop Selector Menu integrated into the main game upon game start.
This is a rough integration, with templates.
That means the functionality is mostly there, but not all the dots are connected!
Also, the TroopTypeSelector was renamed to TroopMetcher. I kept getting confused with the naming of gdscripts, so this might break some things but I think I fixed them all