Troop Deployment
Troops can now be deployed in the game.
Troops cannot be placed on an obstructed tile.
Fixed the music manager not looping audio (for now it always loops)
Changed how troop info is transferred in the drag boxes; instead of confusing arrays for transferred data, they are now just the troop class data transfers.
Made Icons folder, since there will also be portraits and UI and such in the assets later on.
I think a couple other small changes but I been working for a few hours and it is 5:30 AM